+++ src/main/java/de/geheimagentnr1/manyideas_core/elements/blocks/ModBlocks.java
@@ -57,6 +57,7 @@ public class ModBlocks {
 		RegistryEntry.create( PlanksSeamlessCrimson.registry_name, new PlanksSeamlessCrimson() ),//BCPFINRLT
 		RegistryEntry.create( PlanksSeamlessDarkOak.registry_name, new PlanksSeamlessDarkOak() ),//BCPFINRLT
 		RegistryEntry.create( PlanksSeamlessJungle.registry_name, new PlanksSeamlessJungle() ),//BCPFINRLT
+		RegistryEntry.create( PlanksSeamlessMangrove.registry_name, new PlanksSeamlessMangrove() ),//BCPFINRLT
 		RegistryEntry.create( PlanksSeamlessOak.registry_name, new PlanksSeamlessOak() ),//BCPFINRLT
 		RegistryEntry.create( PlanksSeamlessSpruce.registry_name, new PlanksSeamlessSpruce() ),//BCPFINRLT
 		RegistryEntry.create( PlanksSeamlessWarped.registry_name, new PlanksSeamlessWarped() ),//BCPFINRLT
@@ -77,6 +78,7 @@ public class ModBlocks {
 		RegistryEntry.create( LogStrippedSmoothCrimson.registry_name, new LogStrippedSmoothCrimson() ),//BCPFINRLT
 		RegistryEntry.create( LogStrippedSmoothDarkOak.registry_name, new LogStrippedSmoothDarkOak() ),//BCPFINRLT
 		RegistryEntry.create( LogStrippedSmoothJungle.registry_name, new LogStrippedSmoothJungle() ),//BCPFINRLT
+		RegistryEntry.create( LogStrippedSmoothMangrove.registry_name, new LogStrippedSmoothMangrove() ),//BCPFINRLT
 		RegistryEntry.create( LogStrippedSmoothOak.registry_name, new LogStrippedSmoothOak() ),//BCPFINRLT
 		RegistryEntry.create( LogStrippedSmoothSpruce.registry_name, new LogStrippedSmoothSpruce() ),//BCPFINRLT
 		RegistryEntry.create( LogStrippedSmoothWarped.registry_name, new LogStrippedSmoothWarped() ),//BCPFINRLT
@@ -86,6 +88,7 @@ public class ModBlocks {
 		RegistryEntry.create( WoodStrippedSmoothCrimson.registry_name, new WoodStrippedSmoothCrimson() ),//BCPFINRLT
 		RegistryEntry.create( WoodStrippedSmoothDarkOak.registry_name, new WoodStrippedSmoothDarkOak() ),//BCPFINRLT
 		RegistryEntry.create( WoodStrippedSmoothJungle.registry_name, new WoodStrippedSmoothJungle() ),//BCPFINRLT
+		RegistryEntry.create( WoodStrippedSmoothMangrove.registry_name, new WoodStrippedSmoothMangrove() ),//BCPFINRLT
 		RegistryEntry.create( WoodStrippedSmoothOak.registry_name, new WoodStrippedSmoothOak() ),//BCPFINRLT
 		RegistryEntry.create( WoodStrippedSmoothSpruce.registry_name, new WoodStrippedSmoothSpruce() ),//BCPFINRLT
 		RegistryEntry.create( WoodStrippedSmoothWarped.registry_name, new WoodStrippedSmoothWarped() ),//BCPFINRLT
@@ -136,6 +139,10 @@ public class ModBlocks {
 	@ObjectHolder( registryName = RegistryKeys.BLOCKS,
 		value = ManyIdeasCore.MODID + ":" + PlanksSeamlessJungle.registry_name )
 	public static PlanksSeamlessJungle PLANKS_SEAMLESS_JUNGLE;
+	
+	@ObjectHolder( registryName = RegistryKeys.BLOCKS,
+		value = ManyIdeasCore.MODID + ":" + PlanksSeamlessMangrove.registry_name )
+	public static PlanksSeamlessMangrove PLANKS_SEAMLESS_MANGROVE;
 	
 	@ObjectHolder( registryName = RegistryKeys.BLOCKS,
 		value = ManyIdeasCore.MODID + ":" + PlanksSeamlessOak.registry_name )
@@ -210,6 +217,10 @@ public class ModBlocks {
 	public static LogStrippedSmoothJungle LOG_STRIPPED_SMOOTH_JUNGLE;
 	
 	@ObjectHolder( registryName = RegistryKeys.BLOCKS,
+		value = ManyIdeasCore.MODID + ":" + LogStrippedSmoothMangrove.registry_name )
+	public static LogStrippedSmoothMangrove LOG_STRIPPED_SMOOTH_MANGROVE;
+	
+	@ObjectHolder( registryName = RegistryKeys.BLOCKS,
 		value = ManyIdeasCore.MODID + ":" + LogStrippedSmoothOak.registry_name )
 	public static LogStrippedSmoothOak LOG_STRIPPED_SMOOTH_OAK;
 	
@@ -242,6 +253,10 @@ public class ModBlocks {
 	@ObjectHolder( registryName = RegistryKeys.BLOCKS,
 		value = ManyIdeasCore.MODID + ":" + WoodStrippedSmoothJungle.registry_name )
 	public static WoodStrippedSmoothJungle WOOD_STRIPPED_SMOOTH_JUNGLE;
+	
+	@ObjectHolder( registryName = RegistryKeys.BLOCKS,
+		value = ManyIdeasCore.MODID + ":" + WoodStrippedSmoothMangrove.registry_name )
+	public static WoodStrippedSmoothMangrove WOOD_STRIPPED_SMOOTH_MANGROVE;
 	
 	@ObjectHolder( registryName = RegistryKeys.BLOCKS,
 		value = ManyIdeasCore.MODID + ":" + WoodStrippedSmoothOak.registry_name )