+++ src/main/java/de/geheimagentnr1/manyideas_core/elements/items/panels/PanelMangrove.java
@@ -1,0 +1,15 @@
+package de.geheimagentnr1.manyideas_core.elements.items.panels;
+
+import net.minecraft.world.item.Item;
+
+
+public class PanelMangrove extends Item {
+	
+	
+	public static final String registry_name = "panel_mangrove";
+	
+	public PanelMangrove() {
+		
+		super( new Properties() );
+	}
+}